requirements.txt
Flask
google-genai
flask-login
bcrypt
Flask-Limiter
python-dotenv 